executor: add some missing test suite back
