+++ executor/seqtest/seq_executor_test.go
@@ -25,6 +25,7 @@ import (
 	"runtime/pprof"
 	"strconv"
 	"strings"
+	"sync"
 	"sync/atomic"
 	"testing"
 	"time"
@@ -62,6 +63,7 @@ func TestT(t *testing.T) {
 }
 
 var _ = Suite(&seqTestSuite{})
+var _ = Suite(&seqTestSuite1{})
 
 type seqTestSuite struct {
 	cluster   *mocktikv.Cluster
@@ -787,9 +789,130 @@ func (s *seqTestSuite) TestCartesianProduct(c *C) {
 	plannercore.AllowCartesianProduct.Store(true)
 }
 
+func (s *seqTestSuite) TestBatchInsertDelete(c *C) {
+	originLimit := atomic.LoadUint64(&kv.TxnEntryCountLimit)
+	defer func() {
+		atomic.StoreUint64(&kv.TxnEntryCountLimit, originLimit)
+	}()
+	// Set the limitation to a small value, make it easier to reach the limitation.
+	atomic.StoreUint64(&kv.TxnEntryCountLimit, 100)
+
+	tk := testkit.NewTestKit(c, s.store)
+	tk.MustExec("use test")
+	tk.MustExec("drop table if exists batch_insert")
+	tk.MustExec("create table batch_insert (c int)")
+	tk.MustExec("drop table if exists batch_insert_on_duplicate")
+	tk.MustExec("create table batch_insert_on_duplicate (id int primary key, c int)")
+	// Insert 10 rows.
+	tk.MustExec("insert into batch_insert values (1),(1),(1),(1),(1),(1),(1),(1),(1),(1)")
+	r := tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("10"))
+	// Insert 10 rows.
+	tk.MustExec("insert into batch_insert (c) select * from batch_insert;")
+	r = tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("20"))
+	// Insert 20 rows.
+	tk.MustExec("insert into batch_insert (c) select * from batch_insert;")
+	r = tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("40"))
+	// Insert 40 rows.
+	tk.MustExec("insert into batch_insert (c) select * from batch_insert;")
+	r = tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("80"))
+	// Insert 80 rows.
+	tk.MustExec("insert into batch_insert (c) select * from batch_insert;")
+	r = tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("160"))
+	// for on duplicate key
+	for i := 0; i < 160; i++ {
+		tk.MustExec(fmt.Sprintf("insert into batch_insert_on_duplicate values(%d, %d);", i, i))
+	}
+	r = tk.MustQuery("select count(*) from batch_insert_on_duplicate;")
+	r.Check(testkit.Rows("160"))
+
+	// This will meet txn too large error.
+	_, err := tk.Exec("insert into batch_insert (c) select * from batch_insert;")
+	c.Assert(err, NotNil)
+	c.Assert(kv.ErrTxnTooLarge.Equal(err), IsTrue)
+	r = tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("160"))
+
+	// for on duplicate key
+	_, err = tk.Exec(`insert into batch_insert_on_duplicate select * from batch_insert_on_duplicate as tt
+		on duplicate key update batch_insert_on_duplicate.id=batch_insert_on_duplicate.id+1000;`)
+	c.Assert(err, NotNil)
+	c.Assert(kv.ErrTxnTooLarge.Equal(err), IsTrue, Commentf("%v", err))
+	r = tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("160"))
+
+	// Change to batch inset mode and batch size to 50.
+	tk.MustExec("set @@session.tidb_batch_insert=1;")
+	tk.MustExec("set @@session.tidb_dml_batch_size=50;")
+	tk.MustExec("insert into batch_insert (c) select * from batch_insert;")
+	r = tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("320"))
+
+	// Enlarge the batch size to 150 which is larger than the txn limitation (100).
+	// So the insert will meet error.
+	tk.MustExec("set @@session.tidb_dml_batch_size=150;")
+	_, err = tk.Exec("insert into batch_insert (c) select * from batch_insert;")
+	c.Assert(err, NotNil)
+	c.Assert(kv.ErrTxnTooLarge.Equal(err), IsTrue)
+	r = tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("320"))
+	// Set it back to 50.
+	tk.MustExec("set @@session.tidb_dml_batch_size=50;")
+
+	// for on duplicate key
+	_, err = tk.Exec(`insert into batch_insert_on_duplicate select * from batch_insert_on_duplicate as tt
+		on duplicate key update batch_insert_on_duplicate.id=batch_insert_on_duplicate.id+1000;`)
+	c.Assert(err, IsNil)
+	r = tk.MustQuery("select count(*) from batch_insert_on_duplicate;")
+	r.Check(testkit.Rows("160"))
+
+	// Disable BachInsert mode in transition.
+	tk.MustExec("begin;")
+	_, err = tk.Exec("insert into batch_insert (c) select * from batch_insert;")
+	c.Assert(err, NotNil)
+	c.Assert(kv.ErrTxnTooLarge.Equal(err), IsTrue)
+	tk.MustExec("rollback;")
+	r = tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("320"))
+
+	tk.MustExec("drop table if exists com_batch_insert")
+	tk.MustExec("create table com_batch_insert (c int)")
+	sql := "insert into com_batch_insert values "
+	values := make([]string, 0, 200)
+	for i := 0; i < 200; i++ {
+		values = append(values, "(1)")
+	}
+	sql = sql + strings.Join(values, ",")
+	tk.MustExec(sql)
+	tk.MustQuery("select count(*) from com_batch_insert;").Check(testkit.Rows("200"))
+
+	// Test case for batch delete.
+	// This will meet txn too large error.
+	_, err = tk.Exec("delete from batch_insert;")
+	c.Assert(err, NotNil)
+	c.Assert(kv.ErrTxnTooLarge.Equal(err), IsTrue)
+	r = tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("320"))
+	// Enable batch delete and set batch size to 50.
+	tk.MustExec("set @@session.tidb_batch_delete=on;")
+	tk.MustExec("set @@session.tidb_dml_batch_size=50;")
+	tk.MustExec("delete from batch_insert;")
+	// Make sure that all rows are gone.
+	r = tk.MustQuery("select count(*) from batch_insert;")
+	r.Check(testkit.Rows("0"))
+}
+
 type checkPrioClient struct {
 	tikv.Client
 	priority pb.CommandPri
+	mu       struct {
+		sync.RWMutex
+		checkPrio bool
+	}
 }
 
 func (c *checkPrioClient) setCheckPriority(priority pb.CommandPri) {
@@ -802,10 +925,16 @@ func (c *checkPrioClient) getCheckPriority() pb.CommandPri {
 
 func (c *checkPrioClient) SendRequest(ctx context.Context, addr string, req *tikvrpc.Request, timeout time.Duration) (*tikvrpc.Response, error) {
 	resp, err := c.Client.SendRequest(ctx, addr, req, timeout)
-	switch req.Type {
-	case tikvrpc.CmdCop:
-		if c.getCheckPriority() != req.Priority {
-			return nil, errors.New("fail to set priority")
+	c.mu.RLock()
+	defer func() {
+		c.mu.RUnlock()
+	}()
+	if c.mu.checkPrio {
+		switch req.Type {
+		case tikvrpc.CmdCop:
+			if c.getCheckPriority() != req.Priority {
+				return nil, errors.New("fail to set priority")
+			}
 		}
 	}
 	return resp, err
@@ -854,6 +983,10 @@ func (s *seqTestSuite1) TestCoprocessorPriority(c *C) {
 	}
 
 	cli := s.cli
+	cli.mu.Lock()
+	cli.mu.checkPrio = true
+	cli.mu.Unlock()
+
 	cli.setCheckPriority(pb.CommandPri_High)
 	tk.MustQuery("select id from t where id = 1")
 	tk.MustQuery("select * from t1 where id = 1")
@@ -889,4 +1022,8 @@ func (s *seqTestSuite1) TestCoprocessorPriority(c *C) {
 
 	cli.setCheckPriority(pb.CommandPri_Low)
 	tk.MustQuery("select LOW_PRIORITY id from t where id = 1")
+
+	cli.mu.Lock()
+	cli.mu.checkPrio = false
+	cli.mu.Unlock()
 }
